+++ src/main/java/project/web/mvc/repository/AcademyRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import project.web.mvc.domain.Academy;
 
-public interface AcademyRepository extends CrudRepository<Academy, Integer> {
+public interface AcademyRepository extends CrudRepository<Academy, Long> {
 
 }
