+++ src/main/java/project/web/mvc/repository/CartRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import project.web.mvc.domain.Cart;
 
-public interface CartRepository extends CrudRepository<Cart, Integer> {
+public interface CartRepository extends CrudRepository<Cart, Long> {
 
 }
